+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
 ## Unreleased
-- /
+
+- revert fix for ignoring commented properties
+- introduced error for commented properties due to conflicts with Gradle property expansion
 
 ## [1.5.1] - 2026-03-24